% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eaf.R
\name{check_eaf}
\alias{check_eaf}
\title{Check EAF for Exposure and Outcome}
\usage{
check_eaf(dat, pop = "EUR")
}
\arguments{
\item{dat}{A harmonized dataset containing the genetic summary statistics.}

\item{pop}{Population for reference panel, default is "EUR".}
}
\value{
The input dataset with EAF information checked for both exposure and outcome. Messages are displayed for missing values.
}
\description{
This function checks the effect allele frequency (EAF) in both exposure and outcome data
within a harmonized dataset. EAF cannot be added using reID because reID is only available in the reference panel.
The effect allele (EA) in the reference panel may not necessarily be the EA for the reID in your dataset.
This can only be addressed by comparing the EA and non-effect allele (NEA) in the reference panel
versus the EA and NEA in your own dataset.
}
\examples{
# Assuming 'dat' is a harmonized dataset
check_eaf(dat, pop = "EUR")

}
