% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ea_ref_ncb.R
\name{find_ea_ref_ncb}
\alias{find_ea_ref_ncb}
\title{find_ea_ref_ncb}
\usage{
find_ea_ref_ncb(df, pop = "European")
}
\arguments{
\item{dat}{A dataframe containing MAF columns.}
}
\value{
A dataframe with ea_ref column from 1000G.
}
\description{
Compare your maf with 1000G as reference panel.
}
\examples{
dat <- find_ea_ref_ncb(dat)


}
