% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_res.R
\name{log_res}
\alias{log_res}
\title{per doubling in genetic liability for xxx}
\usage{
log_res(res)
}
\arguments{
\item{res}{}
}
\value{
log res
}
\description{
https://arxiv.org/abs/1804.05545
}
\examples{
res <- log_res(res)


}
