% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U1_maf2eaf.R
\name{U1_maf2eaf}
\alias{U1_maf2eaf}
\title{U1_maf2eaf}
\usage{
U1_maf2eaf(df)
}
\arguments{
\item{df}{A dataframe containing MAF or EAF columns.}
}
\value{
A dataframe with EAF column, preserving the original MAF column.
}
\description{
Convert MAF to EAF in a dataframe, checking for and renaming relevant columns.
}
\examples{
df <- U1_maf2eaf(df)


}
