% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U4_check_samplesize_rerun.R
\name{U4_check_samplesize_rerun}
\alias{U4_check_samplesize_rerun}
\title{check samplesize}
\usage{
U4_check_samplesize_rerun(dat)
}
\arguments{
\item{dat}{dat after U4_harmonise_data}
}
\value{
dat
}
\description{
Check if the sample size is correct, if not, it needs to be modified. If the name is not satisfactory, it also needs to be modified. Do remember rerun again after correction.
}
\examples{

help(U4_harmonise_data)



}
