% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U4_harmonisation.R
\name{U4_harmonisation}
\alias{U4_harmonisation}
\title{U4_harmonisation merges exposure and outcome data into dat}
\usage{
U4_harmonisation(exposure_dat, outcome_dat, action = 2)
}
\arguments{
\item{exposure_dat}{Instrumental variables extracted by U2_extract_instruments}

\item{outcome_dat}{Outcomes extracted by U3_extract_outcomes_data}

\item{action}{Default is 2, refer to the description of harmonise_data in TwoSampleMR}
}
\value{
Harmonised data
}
\description{
U4_harmonisation merges exposure and outcome data into dat
}
\examples{
dat <- U4_harmonisation(IVs, Outs)

\dontrun{
  IVs <- U2_extract_instruments(c("ieu-a-31", "ieu-a-300", "ieu-a-2"))
  Outs <- U3_extract_outcomes_data(
    outcome = c("ieu-a-297", "ieu-a-298", "ieu-a-299"),
    exposure_iv = IVs
  )
  dat <- U4_harmonisation(IVs, Outs) \%>\%
    U4_add_eaf() \%>\%
    U4_check_samplesize()


#### Check if the sample size is correct ######################################

# If there are too many IDs and they are not fully displayed, you can use View(dat) to check "description_exposure" and "description_outcome"

#### Exposure part ###########################################################

# If the sample size of Inflammatory bowel disease is incorrect, it needs to be modified

# Modify total sample size
dat$samplesize.exposure[which(dat$exposure == "Inflammatory bowel disease")] <- 12000

# Modify ncase
dat$ncase.exposure[which(dat$exposure == "Inflammatory bowel disease")] <- 1200

# Modify ncontrol
dat$ncontrol.exposure[which(dat$exposure == "Inflammatory bowel disease")] <- 10800

# Modify name
dat$exposure[which(dat$exposure == "Inflammatory bowel disease")] <- "IBD"

#### Outcome part ############################################################

# If HDL cholesterol is actually a binary variable, it needs to be modified

# Add ncase
dat$ncase.outcome[which(dat$outcome == "HDL cholesterol")] <- 1200
# Add ncontrol
dat$ncontrol.outcome[which(dat$outcome == "HDL cholesterol")] <- 10800


# For the outcome name Alzheimer's disease, if you want to modify the first Alzheimer's disease

dat$outcome[which(dat$id.outcome == "ieu-a-297")] <- "AD"


#### Final confirmation ######################################################

dat <- U4_check_samplesize(dat)

###### If you want to remove SNPs with F value less than 10 ##################

dat <- subset(dat, singleF > 10)
print(subset(dat, singleF <= 10)$SNP)

###### steiger_filtering #####################################################

dat <- dat \%>\%
  TwoSampleMR::steiger_filtering() \%>\%
  {removed_snps <- dplyr::filter(., steiger_dir != TRUE) \%>\% dplyr::pull(SNP);
  cat("Removed SNPs:\n", removed_snps, "\n\nTotal number of removed SNPs:", length(removed_snps), "\n");
  dplyr::filter(., steiger_dir == TRUE)}

}
}
