% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U6_mrcML.R
\name{U6_mrcML}
\alias{U6_mrcML}
\title{U6_mrcML}
\usage{
U6_mrcML(
  dat,
  res,
  random_seed = 314,
  Alpha = 0.05,
  num_pert = 200,
  maxit = 100
)
}
\arguments{
\item{dat}{dat after harmonisation}

\item{res}{res after mr}
}
\value{
mrcML_res
}
\description{
further analysis for univariable cML from MendelianRandomization package add to res
}
\examples{
mrcML_res <- U6_mrcML(dat, res, random_seed = 314, Alpha = 0.05, num_pert = 200, maxit = 100)


}
