% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_to_mr_input.R
\name{dat_to_mr_input}
\alias{dat_to_mr_input}
\title{dat_to_mr_input}
\usage{
dat_to_mr_input(dat)
}
\arguments{
\item{dat}{}
}
\value{
mrinput
}
\description{
mrinput from MendelianRandomization package
}
\examples{
mr_input_obj <- dat_to_mr_input(dat)


}
