% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U1_Clean_data_exp.R
\name{U1_Clean_data_exp}
\alias{U1_Clean_data_exp}
\title{U1_Clean_data_exp}
\usage{
U1_Clean_data_exp(df, continuous = FALSE, save = FALSE)
}
\arguments{
\item{df}{Dataframe}

\item{continuous}{If TRUE, will not ask for case sample size.}

\item{save}{Default is FALSE. If TRUE, will save the dataframe as a local file with the name of the id.}
}
\value{
Returns the cleaned data.
}
\description{
Clean exposure data. Exposure does not require many SNPs. https://gitee.com/OneClickAnalyses/Oneclick/blob/master/R/U1_Clean_data.R
}
\examples{
df <- U1_Clean_data_exp(df)

}
