% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U6_mrcML.R
\name{U6_mrcML}
\alias{U6_mrcML}
\title{U6_mrcML}
\usage{
U6_mrcML(
  mr_input_obj,
  dat,
  random_seed = 314,
  Alpha = 0.05,
  num_pert = 200,
  maxit = 100
)
}
\arguments{
\item{dat}{}
}
\value{
mrcML_res
}
\description{
further analysis for univariable cML from MendelianRandomization package
}
\examples{
mrcML_res <- dat \%>\%
 dat_to_mr_input() \%>\%
 U6_mrcML(dat = ., random_seed = 314, Alpha = 0.05, num_pert = 200, maxit = 100)


}
